refactor: queries와 mutations 분리
