+++ packages/shared/src/hooks/supabase/index.ts
@@ -1,14 +1,2 @@
-export { default as useCategoryDeleteMutation } from './useCategoryDeleteMutation';
-export { default as useCategoryPatchMutation } from './useCategoryPatchMutation';
-export { default as useCategoryPostMutation } from './useCategoryPostMutation';
-export { default as useCategoryQuery } from './useCategoryQuery';
-export { default as useCategoryUpsertMutation } from './useCategoryUpsertMutation';
-export { default as useDeleteMemosMutation } from './useDeleteMemosMutation';
-export { default as useMemoPatchMutation } from './useMemoPatchMutation';
-export { default as useMemoPostMutation } from './useMemoPostMutation';
-export { default as useMemoQuery } from './useMemoQuery';
-export { default as useMemosQuery } from './useMemosQuery';
-export { default as useMemosUpsertMutation } from './useMemosUpsertMutation';
-export { default as useSupabaseUser } from './useSupabaseUser';
-export { default as useSignoutMutation } from './useSignoutMutation';
-export { default as useFeedbackMutation } from './useFeedbackMutation';
+export * from './queries';
+export * from './mutations';
