+++ packages/shared/src/hooks/supabase/mutations/useCategoryDeleteMutation.ts
@@ -2,14 +2,14 @@ import { QUERY_KEY } from '@src/constants';
 import { CategoryService } from '@src/utils';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
-import useSupabaseClientQuery from './useSupabaseClientQuery';
+import { useSupabaseClientQuery } from '../queries';
 
-export default function useCategoryPostMutation() {
+export default function useCategoryDeleteMutation() {
   const queryClient = useQueryClient();
   const { data: supabaseClient } = useSupabaseClientQuery();
 
   return useMutation({
-    mutationFn: new CategoryService(supabaseClient).insertCategory,
+    mutationFn: new CategoryService(supabaseClient).deleteCategory,
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: QUERY_KEY.category() });
     },
