+++ packages/shared/src/hooks/supabase/mutations/useDeleteMemosMutation.ts
@@ -1,32 +1,27 @@
 import { NoMemosError, QUERY_KEY } from '@src/constants';
-import type { MemoSupabaseResponse, MemoTable } from '@src/types';
+import { MemoSupabaseResponse } from '@src/types';
 import { MemoService } from '@src/utils';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
-import useSupabaseClientQuery from './useSupabaseClientQuery';
+import { useSupabaseClientQuery } from '../queries';
 
-type MutationError = Error;
-
-export default function useMemoPostMutation() {
+export default function useDeleteMemosMutation() {
   const queryClient = useQueryClient();
   const { data: supabaseClient } = useSupabaseClientQuery();
 
-  return useMutation<MemoSupabaseResponse, MutationError, MemoTable['Insert']>({
-    mutationFn: new MemoService(supabaseClient).insertMemo,
-    onSuccess: async result => {
-      const { data: newData } = result;
-
+  return useMutation<MemoSupabaseResponse, Error, number[]>({
+    mutationFn: new MemoService(supabaseClient).deleteMemos,
+    onMutate: async idList => {
       await queryClient.cancelQueries({ queryKey: QUERY_KEY.memos() });
-
       const previousMemos = queryClient.getQueryData<MemoSupabaseResponse>(QUERY_KEY.memos());
 
-      if (!previousMemos || !newData) throw new NoMemosError();
+      if (!previousMemos) throw new NoMemosError();
 
       const { data: previousMemosData } = previousMemos;
 
       if (!previousMemosData) throw new NoMemosError();
 
-      const newMemosData = newData.concat(previousMemosData);
+      const newMemosData = previousMemosData.filter(memo => !idList.includes(memo.id));
 
       await queryClient.setQueryData(QUERY_KEY.memos(), { ...previousMemos, data: newMemosData });
