+++ packages/shared/src/hooks/supabase/mutations/useMemoPostMutation.ts
@@ -1,27 +1,32 @@
 import { NoMemosError, QUERY_KEY } from '@src/constants';
-import { MemoSupabaseResponse } from '@src/types';
+import type { MemoSupabaseResponse, MemoTable } from '@src/types';
 import { MemoService } from '@src/utils';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
-import useSupabaseClientQuery from './useSupabaseClientQuery';
+import { useSupabaseClientQuery } from '../queries';
 
-export default function useDeleteMemosMutation() {
+type MutationError = Error;
+
+export default function useMemoPostMutation() {
   const queryClient = useQueryClient();
   const { data: supabaseClient } = useSupabaseClientQuery();
 
-  return useMutation<MemoSupabaseResponse, Error, number[]>({
-    mutationFn: new MemoService(supabaseClient).deleteMemos,
-    onMutate: async idList => {
+  return useMutation<MemoSupabaseResponse, MutationError, MemoTable['Insert']>({
+    mutationFn: new MemoService(supabaseClient).insertMemo,
+    onSuccess: async result => {
+      const { data: newData } = result;
+
       await queryClient.cancelQueries({ queryKey: QUERY_KEY.memos() });
+
       const previousMemos = queryClient.getQueryData<MemoSupabaseResponse>(QUERY_KEY.memos());
 
-      if (!previousMemos) throw new NoMemosError();
+      if (!previousMemos || !newData) throw new NoMemosError();
 
       const { data: previousMemosData } = previousMemos;
 
       if (!previousMemosData) throw new NoMemosError();
 
-      const newMemosData = previousMemosData.filter(memo => !idList.includes(memo.id));
+      const newMemosData = newData.concat(previousMemosData);
 
       await queryClient.setQueryData(QUERY_KEY.memos(), { ...previousMemos, data: newMemosData });
