+++ packages/shared/src/hooks/supabase/mutations/useSignoutMutation.ts
@@ -1,5 +1,5 @@
 import { useMutation } from '@tanstack/react-query';
-import useSupabaseClientQuery from './useSupabaseClientQuery';
+import { useSupabaseClientQuery } from '../queries';
 import { AuthService } from '@src/utils/Supabase';
 
 export default function useSignoutMutation() {
