+++ pages/side-panel/src/hoc/withAuthentication.tsx
@@ -1,9 +1,9 @@
-import { useSupabaseUser } from '@extension/shared/hooks';
+import { useSupabaseUserQuery } from '@extension/shared/hooks';
 import { LoginSection } from '@src/components';
 
 export default function withAuthentication(WrappedComponent: () => JSX.Element) {
   const AuthenticatedComponent = function () {
-    const { user } = useSupabaseUser();
+    const { user } = useSupabaseUserQuery();
 
     if (user?.data.user) return <WrappedComponent />;
     return <LoginSection />;
